requirements.txt
absl-py==2.0.0
altair==5.4.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
babel==2.16.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.1.14
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
importlib_metadata==8.4.0
incremental==22.10.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.7
mistune==3.0.2
modules==1.0.0
narwhals==1.10.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.24.2
opencv-contrib-python==4.8.1.78
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.1
pandas==2.0.0
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.4
Pillow==9.5.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
Protego==0.2.1
protobuf==3.20.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydeck==0.9.1
PyDispatcher==2.0.7
pygame==2.6.0
Pygments==2.18.0
pyodbc==4.0.35
pyOpenSSL==23.1.1
pyparsing==3.0.9
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytube==15.0.0
pytz==2023.3
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.6.2
referencing==0.35.1
requests==2.32.3
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.13.1
Scrapy==2.8.0
selenium==4.15.2
Send2Trash==1.8.3
service-identity==21.1.0
shiboken6==6.5.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.4.1
SpeechRecognition==3.10.0
stack-data==0.6.3
streamlit==1.39.0
streamlit-elements==0.1.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.10.0
twisted-iocpsupport==1.0.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.15
w3lib==2.1.1
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wsproto==1.2.0
zipp==3.15.0
zope.interface==6.0
